modulo 3 - aula 03

+++ app/src/main/java/br/com/alura/agenda/dao/AlunoDAO.java
@@ -18,7 +18,7 @@ import br.com.alura.agenda.modelo.Aluno;
  */
 public class AlunoDAO extends SQLiteOpenHelper {
     public AlunoDAO(Context context) {
-        super(context, "Agenda", null, 5);
+        super(context, "Agenda", null, 6);
     }
 
     @Override
@@ -31,7 +31,8 @@ public class AlunoDAO extends SQLiteOpenHelper {
                 "site TEXT, " +
                 "nota REAL, " +
                 "caminhoFoto TEXT," +
-                "sincronizado INT DEFAULT 0);";
+                "sincronizado INT DEFAULT 0," +
+                "desativado INT DEFAULT 0);";
         db.execSQL(sql);
     }
 
@@ -76,6 +77,10 @@ public class AlunoDAO extends SQLiteOpenHelper {
             case 4:
                 String adicionaCampoSincronizado = "ALTER TABLE Alunos ADD COLUMN sincronizado INT DEFAULT 0";
                 db.execSQL(adicionaCampoSincronizado);
+            case 5:
+                String adicionaCampoDesativado =
+                        "ALTER TABLE Alunos ADD COLUMN desativado INT DEFAULT 0;";
+                db.execSQL(adicionaCampoDesativado);
 
         }
 
@@ -140,11 +145,12 @@ public class AlunoDAO extends SQLiteOpenHelper {
         dados.put("nota", aluno.getNota());
         dados.put("caminhoFoto", aluno.getCaminhoFoto());
         dados.put("sincronizado", aluno.getSincronizado());
+        dados.put("desativado", aluno.getDesativado());
         return dados;
     }
 
     public List<Aluno> buscaAlunos() {
-        String sql = "SELECT * FROM Alunos;";
+        String sql = "SELECT * FROM Alunos WHERE desativado = 0;";
         SQLiteDatabase db = getReadableDatabase();
         Cursor c = db.rawQuery(sql, null);
 
@@ -168,6 +174,7 @@ public class AlunoDAO extends SQLiteOpenHelper {
             aluno.setNota(c.getDouble(c.getColumnIndex("nota")));
             aluno.setCaminhoFoto(c.getString(c.getColumnIndex("caminhoFoto")));
             aluno.setSincronizado(c.getInt(c.getColumnIndex("sincronizado")));
+            aluno.setDesativado(c.getInt(c.getColumnIndex("desativado")));
 
             alunos.add(aluno);
         }
@@ -178,7 +185,12 @@ public class AlunoDAO extends SQLiteOpenHelper {
         SQLiteDatabase db = getWritableDatabase();
 
         String[] params = {aluno.getId().toString()};
-        db.delete("Alunos", "id = ?", params);
+        if (aluno.estaDesativado()){
+            db.delete("Alunos", "id = ?", params);
+        }else {
+            aluno.desativa();
+            altera(aluno);
+        }
     }
 
     public void altera(Aluno aluno) {
